Separate index.ts into editable/editingArea/editorField/labelContainer.ts
